+++ documentation/documentation-default-settings-update.pl
@@ -34,7 +34,7 @@ my @namesAndOffsets = (
                         {name=>"modifylinebreaks",numberOfLines=>2,special=>"modifyLineBreaks",mustBeAtBeginning=>1},
                         {name=>"textWrapOptions",numberOfLines=>1},
                         {name=>"textWrapOptionsAll",numberOfLines=>22,special=>"textWrapOptions"},
-                        {name=>"oneSentencePerLine",numberOfLines=>23},
+                        {name=>"oneSentencePerLine",numberOfLines=>24},
                         {name=>"sentencesFollow",numberOfLines=>8},
                         {name=>"sentencesBeginWith",numberOfLines=>3},
                         {name=>"sentencesEndWith",numberOfLines=>5},
@@ -325,6 +325,8 @@ if(!$readTheDocsMode){
                         $index_body =~ s|!|;|s;
                     }
                     "\n.. index:: ".$index_body."\n\n";/sgex;
+
+        $body =~ s/\.\. index::/\n\n.. index::/sg;
 
         # line numbers for defaulSettings
         for (@namesAndOffsets){