+++ documentation/latex-indent.sh
@@ -1,13 +1,12 @@
 #!/bin/bash
-loopmax=1
 . ../test-cases/common.sh
 
-# if silentMode is not active, verbose
 [[ $silentMode == 0 ]] && set -x 
 
 latexindent.pl -l=+mainfile.yaml -w -rv -m -s  latexindent.tex
+latexindent.pl -w -rv -m -g tex-files.log -l -s s*.tex
+latexindent.pl -w -rv -m -g bib-files.log -l bibsettings.yaml -s *.bib
 
-find . -maxdepth 1 -name "s*.tex" -exec latexindent.pl -l -w -rv -m -s {} \;
-find . -maxdepth 1 -name "*.bib" -exec latexindent.pl -l bibsettings.yaml -w -rv -m -s {} \;
+set +x
 [[ $noisyMode == 1 ]] && makenoise
 [[ $gitStatus == 1 ]] && git status