+++ helper-scripts/update-version.sh
@@ -10,12 +10,13 @@
 #   update-version.sh 
 
 minorVersion=0
-oldVersion='3.15'
-newVersion='3.16'
-oldDate='2022-01-21'
-newDate='2022-03-13'
+oldVersion='3.16'
+newVersion='3.17'
+oldDate='2022-03-13'
+newDate='2022-03-25'
+updateVersion=0
 
-while getopts "hmv" OPTION
+while getopts "hmuv" OPTION
 do
  case $OPTION in 
   h)
@@ -26,6 +27,8 @@ ${0##*/} [OPTIONS]
     -h  help, outputs this message
 
     -m  minor version, not removing most recently updated stars from documentation
+
+    -u  update version
 
 Typical running order pre-release:
 
@@ -48,6 +51,10 @@ ____PLEH
    echo "minor version, not removing most recently updated stars from documentation..."
    minorVersion=1
    ;;
+  u)    
+   echo "update switch active"
+   updateVersion=1
+   ;;
   v)    
    echo "version details:"
    echo "old version details: $oldVersion, $oldDate"
@@ -64,19 +71,23 @@ done
 echo "old version details: $oldVersion, $oldDate"
 echo "NEW version details: $newVersion, $newDate"
 
+[[ $updateVersion == 0 ]] && printf "not updating, you can run the following instead\n\n    update-version.sh -u\n\n" && exit 0
+
 cd ../
 cd documentation
 [[ $minorVersion == 0 ]] && find -name "s*.tex" -print0|xargs -0 perl -p0i -e "s|announce\*\{|announce\{|sg"
 
 set -x
+
+# change \announce{new}{message} into \announce*{message}
+find -name "s*.tex" -print0|xargs -0 perl -p0i -e "s|announce\{new|announce\*\{$newDate|sgi"
+
 pdflatex --interaction=batchmode latexindent
 
 # update .rst
 perl documentation-default-settings-update.pl
 perl documentation-default-settings-update.pl -r
 
-# change \announce{new}{message} into \announce*{message}
-find -name "s*.tex" -print0|xargs -0 perl -p0i -e "s|announce\{new|announce\*\{$newDate|sgi"
 set +x
 
 cd ../
@@ -111,7 +122,15 @@ done
 # check for new announcements in the documentation
 egrep -i --color=auto 'announce{new' documentation/*.tex
 
+# check for cmhlistings* in the documentation
+# 
+# note:
+#
+#   sed -i.bak "s/cmhlistingsfromfile\*/cmhlistingsfromfile/" $file
+egrep -i --color=auto 'cmhlistingsfromfile\*' documentation/s*.tex
+
 # update changelog.md manually
+cat documentation/latexindent.new
 gedit documentation/changelog.md
 
 cat <<-____TXEN