+++ test-cases/specials/specials-test-cases.sh
@@ -61,8 +61,8 @@ latexindent.pl -s algpseudocode-mk2.tex -o=+-mod-array -l=koppor2
 latexindent.pl -s algpseudocode.tex -o=+-mod3 -l=koppor3
 latexindent.pl -s algpseudocode.tex -o=+-mod4 -l=koppor4
 # specialMiddle polyswitch
-[[ $loopmin -gt 10 ]] && loopmin=1 && echo "loopmin set to 1 (special middle polyswitch testing)"
-[[ $loopmax -gt 10 ]] && loopmax=10 && echo "loopmax set to 10 (special middle polyswitch testing)"
+[[ $loopmin -gt 10 ]] && loopmin=1
+[[ $loopmax -gt 10 ]] && loopmax=10
 for (( i=$loopmin ; i <= $loopmax ; i++ )) 
 do 
    [[ $showCounter == 1 ]] && echo "$i of $loopmax"