+++ documentation/changelog.md
@@ -1,4 +1,13 @@
 # changelog.md
+
+## V3.17, March 25, 2022
+
+* batches of files now supported, [issue 332](https://github.com/cmhughes/latexindent.pl/issues/332)
+* `-wd` switch which means 'overwrite if different'; similar to the `-w` switch, but will only activate 
+the overwrite and back-up file procedure if the indented text is different from the original
+* `multipleSpacesToSingle` feature for `oneSentencePerLine`
+* `--GCString` switch available to *optionally* load the `Unicode::GCString` module, [issue 303](https://github.com/cmhughes/latexindent.pl/issues/303)
+
 ## V3.16, March 13, 2022
 * text wrap overhaul, [issue 346](https://github.com/cmhughes/latexindent.pl/issues/346)
 * support for `tabulararray`, [issue 343](https://github.com/cmhughes/latexindent.pl/issues/343)
@@ -285,3 +294,5 @@ help in the TeXLive implementation
  bundled into TeXLive 2013. Some other updates, too, detailed in
  the documentation. 
 
+## first commit, August 28, 2012
+[first commit](https://github.com/cmhughes/latexindent.pl/commit/ac8a07df62b87a4b663a64125c33bca51e48dffa)